MNT: rename gnss file
