+++ rocketpy/sensors/__init__.py
@@ -1,5 +1,5 @@
 from .accelerometer import Accelerometer
 from .barometer import Barometer
-from .gnss import GnssReceiver
+from .gnss_receiver import GnssReceiver
 from .gyroscope import Gyroscope
 from .sensor import InertialSensor, ScalarSensor, Sensor
